rewrite unit tests and fix mock

+++ internal/service/cron_test.go
@@ -15,7 +15,7 @@ func Test_RunTimerTask(t *testing.T) {
 	dyn := mocks.NewMockDynamoRepository(ctrl)
 	cw := mocks.NewMockCloudWatchRepository(ctrl)
 	svc := NewProcessorService(tl(), dyn, nil, nil, nil, cw)
-	dyn.EXPECT().ScanCount(gomock.Any()).Return(int64(10), nil)
+	dyn.EXPECT().ScanCount(gomock.Any()).Return(int32(10), nil)
 	cw.EXPECT().PutCountMetric(gomock.Any(), "ns", "ItemsCount", float64(10)).Return(nil)
 	if err := svc.RunTimerTask(context.Background(), "ns"); err != nil {
 		t.Fatalf("unexpected: %v", err)
